Fix portfolio holdings registration and refactor utility functions
PRIORITY 1: Fixed the critical portfolio holdings registration issue:
- Replace domain entity Portfolio with SecurityPortfolioManager
- Connect market_order() to portfolio.process_fill() for proper holdings updates
- Add set_holdings() method that properly manages portfolio positions
- Fix engine to create connected order execution pipeline

REFACTORING: Move utility functions to appropriate base classes:
- Move data processing utilities from MyAlgorithm to QCAlgorithm base class:
  * _has_price_data(), _get_current_price(), _get_current_holdings_value()
  * _symbol_in_data(), _find_matching_symbol()
- Move web interface functions from TestProjectBacktestManager to misbuffet:
  * Create new WebInterfaceManager in application.services.misbuffet
  * _setup_progress_logging(), _start_web_interface(), _open_browser()

TECHNICAL IMPROVEMENTS:
- Fix SecurityPortfolioManager integration with proper portfolio access
- Add proper market price lookup with multiple fallback strategies
- Enhanced error handling and logging for portfolio operations
- Clean up imports and remove duplicate functionality
